job/vacancy modified

+++ client/views.py
@@ -22,7 +22,8 @@ from .models import (
     FavouriteStaff,
     MyStaff,
     StaffReview,
-    JobReport
+    JobReport,
+
 
 
 )
@@ -39,7 +40,10 @@ from .serializers import (
     PermanentJobsSerializer,
     FavouriteStaffSerializer,
     MyStaffSerializer,
-    StaffReviewSerializer
+    StaffReviewSerializer,
+
+
+
 
 )
 
@@ -114,74 +118,87 @@ class CompanyProfileCreateView(generics.ListCreateAPIView):
             return Response(response, status=status.HTTP_400_BAD_REQUEST)
 
 class VacancyView(APIView):
-    def get(self, request, pk=None):
-        user = request.user 
-        client = CompanyProfile.objects.filter(user=user).first()
-
+    def get(self, request,job_id=None, pk=None, **kwargs):
+        if job_id:
+            vacancy = Vacancy.objects.filter(job__id=job_id)
+            serializer = VacancySerializer(vacancy, many=True)
+            response = {
+                "status": status.HTTP_200_OK,
+                "success": True,
+                "message": "Vacancies",
+                "data": serializer.data
+            }
+            return Response(response, status=status.HTTP_200_OK)
         if pk:
-            vacancy = Vacancy.objects.filter(client=client, pk=pk).first()
-            if not vacancy:
-                return Response({"error": "Vacancy not found"}, status=status.HTTP_404_NOT_FOUND)
+            vacancy = get_object_or_404(Vacancy, pk=pk)
             serializer = VacancySerializer(vacancy)
-            response_data = {
+            response = {
                 "status": status.HTTP_200_OK,
                 "success": True,
                 "message": "Vacancy details",
                 "data": serializer.data
             }
-            return Response(response_data, status=status.HTTP_200_OK)
+            return Response(response, status=status.HTTP_200_OK)
+        # return error response
+        return Response({"error": "Invalid request"}, status=status.HTTP_400_BAD_REQUEST)
         
-        vacancies = Vacancy.objects.filter(client=client)
-        serializer = VacancySerializer(vacancies, many=True)
-        response_data = {
-            "status": status.HTTP_200_OK,
-            "success": True,
-            "message": "List of vacancies",
-            "data": serializer.data
-        }
-        return Response(response_data, status=status.HTTP_200_OK)
     def post(self, request):
-        data = request.data
-        serializer = CreateVacancySerializers(data=request.data, context={'request': request})
+        
+        serializer = CreateVacancySerializers(data=request.data)
         if serializer.is_valid():
             vacancy = serializer.save()
             response = VacancySerializer(vacancy)
             response_data = {
                 "status": status.HTTP_200_OK,
                 "message": "Vacancy created successfully",
-                # show response data in vacancy serializer
                 "data": response.data
                 
             }
             return Response(response_data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    # update vacancy 
-    def put(self, request, pk):
-        vacancy = Vacancy.objects.get(pk=pk)
-        serializer = CreateVacancySerializers(vacancy, data=request.data, context={'request': request})
+
+    def put(self, request, pk, *args, **kwargs):
+        user = request.user
+        if user.is_client:
+            client = CompanyProfile.objects.filter(user=user).first()
+        else:
+            return Response({"error": "Only client can access this endpoint"}, status=status.HTTP_400_BAD_REQUEST)
+        try:
+            vacancy = Vacancy.objects.get(pk=pk)
+            if vacancy.job.company!= client:
+                return Response({"error": "You are not authorized to access this vacancy"}, status=status.HTTP_403_FORBIDDEN)
+        except Vacancy.DoesNotExist:
+            return Response({"error": "Vacancy not found"}, status=status.HTTP_404_NOT_FOUND)
+        serializer = CreateVacancySerializers(instance=vacancy, data=request.data)
         if serializer.is_valid():
-            res = serializer.save()
-            response = VacancySerializer(res)
-            response = {
+            serializer.save()
+            response = VacancySerializer(vacancy)
+            response_data = {
                 "status": status.HTTP_200_OK,
                 "message": "Vacancy updated successfully",
                 "data": response.data
-                
             }
-            return Response(response, status=status.HTTP_200_OK)
+            return Response(response_data, status=status.HTTP_200_OK)
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    
-    def delete(self, request, pk):
-        job = get_object_or_404(Vacancy, pk=pk)
-        job.delete()
-        response = {
-            "status": status.HTTP_204_NO_CONTENT,
-            "success": True,
-            "message": "Vacancy deleted successfully"
-        }
-        return Response(response, status=status.HTTP_204_NO_CONTENT)
-    
+        
+    def delete(self, request, pk, *args, **kwargs):
+        user = request.user
+        if user.is_client:
+            client = CompanyProfile.objects.filter(user=user).first()
+            if not client:
+                return Response({"error": "Company profile not found"}, status=status.HTTP_404_NOT_FOUND)
+            vacancy = Vacancy.objects.filter(pk=pk, job__company=client).first()
+            if not vacancy:
+                return Response({"error": "You are not authorized to delete this vacancy"}, status=status.HTTP_403_FORBIDDEN)
+            vacancy.delete()
+            response_data = {
+                "status": status.HTTP_200_OK,
+                "success": True,
+                "message": "Vacancy deleted successfully"
+            }
+            return Response(response_data, status=status.HTTP_200_OK)
+
 class JobView(APIView):
     def get(self, request, pk=None, *args, **kwargs):
         user=request.user
@@ -237,24 +254,24 @@ class JobView(APIView):
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
     def delete(self, request, pk):
-        try:
-            user = request.user
+        user = request.user
+        if user.is_client:
             client = CompanyProfile.objects.filter(user=user).first()
-            job = Job.objects.get(pk=pk,company=client)
-        except Job.DoesNotExist:
-            return Response({'error': 'Job not found'}, status=status.HTTP_404_NOT_FOUND)
-        # also delete all the vacancy include with this job
-        for vacancy in job.vacancy.all():
-            vacancy.delete()
-        # delete job itself
-        job.delete()
-        response = {
-            "status": status.HTTP_204_NO_CONTENT,
-            "success": True,
-            "message": "Job deleted successfully"
-        }
-        return Response(response, status=status.HTTP_204_NO_CONTENT)
-    
+            if not client:
+                return Response({"error": "Company profile not found"}, status=status.HTTP_404_NOT_FOUND)
+            job = Job.objects.filter(pk=pk).first()
+            if not job:
+                return Response({"error": "You are not authorized to delete this job"}, status=status.HTTP_403_FORBIDDEN)
+            if job.company == client:
+                job.delete()
+                response_data = {
+                    "status": status.HTTP_200_OK,
+                    "success": True,
+                    "message": "Job deleted successfully"
+                }
+                return Response(response_data, status=status.HTTP_200_OK)
+            return Response({"error": "You are not authorized to delete this job"}, status=status.HTTP_403_FORBIDDEN)
+        
 
 class JobApplicationAPI(APIView): # pending actions page approve job
     def get(self, request, vacancy_id=None,pk=None):
